% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_track.R
\name{plot_track}
\alias{plot_track}
\title{Plot Bird Track}
\usage{
plot_track(
  results,
  colony_lat,
  colony_lon,
  bird_id,
  output_dir,
  hemisphere = ""
)
}
\arguments{
\item{results}{data.frame with Latitude and Longitude columns}

\item{colony_lat}{Numeric latitude of colony}

\item{colony_lon}{Numeric longitude of colony}

\item{bird_id}{Character ID for plot title}

\item{output_dir}{Directory to save plot}

\item{hemisphere}{Character hemisphere check result}
}
\value{
NULL (creates PNG file)
}
\description{
Creates a map showing the estimated movement track of a bird
}
