% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calibrate_gls_batch.R
\name{process_single_bird}
\alias{process_single_bird}
\title{Process a Single Bird}
\usage{
process_single_bird(
  lux_file,
  bird_id,
  colony_lat,
  colony_lon,
  light_threshold,
  exclude_equinoxes,
  output_dir,
  fig_dir,
  verbose
)
}
\description{
Internal function to process a single GLS device
}
\keyword{internal}
