% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gls.R
\name{read_act_file}
\alias{read_act_file}
\title{Read a Migrate Technology wet/dry activity (.act) file}
\usage{
read_act_file(file_path)
}
\arguments{
\item{file_path}{Path to a \code{.act} file.}
}
\value{
data.frame(Date, wet) where \code{wet} is the conductivity count.
}
\description{
Read a Migrate Technology wet/dry activity (.act) file
}
