% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gls.R
\name{read_deg_file}
\alias{read_deg_file}
\title{Read a Migrate Technology temperature (.deg / .tem) file}
\usage{
read_deg_file(file_path)
}
\arguments{
\item{file_path}{Path to a \code{.deg} or \code{.tem} file.}
}
\value{
data.frame(Date, Temp) with temperature in degrees Celsius.
}
\description{
Read a Migrate Technology temperature (.deg / .tem) file
}
